+++ src/bagels/modals/record.py
@@ -10,11 +10,13 @@ from textual.widgets import (
 from bagels.components.autocomplete import AutoComplete, Dropdown, DropdownItem
 from bagels.components.fields import Fields
 from bagels.config import CONFIG
+from bagels.managers.record_templates import get_template_by_id, create_template
 from bagels.modals.input import InputModal
 from bagels.managers.accounts import get_all_accounts_with_balance
 from bagels.managers.persons import create_person, get_all_persons
 from bagels.forms.form import Form, Option
 from bagels.forms.record_forms import RecordForm
+from bagels.utils.format import format_date_to_readable
 from bagels.utils.validation import validateForm
 from bagels.modals.base_widget import ModalContainer
 from datetime import datetime
@@ -28,20 +30,23 @@ class RecordModal(InputModal):
         Binding(
             CONFIG.hotkeys.record_modal.new_split,
             "add_split",
-            "Add split",
+            "+split",
             priority=True,
         ),
         Binding(
             CONFIG.hotkeys.record_modal.new_paid_split,
             "add_paid_split",
-            "Add paid split",
+            "+paid split",
             priority=True,
         ),
         Binding(
             CONFIG.hotkeys.record_modal.delete_last_split,
             "delete_last_split",
-            "Delete last split",
+            "-last split",
             priority=True,
+        ),
+        Binding(
+            "shift+enter", "submit_and_template", "Submit & Template", priority=True
         ),
     ]
 
@@ -67,6 +72,7 @@ class RecordModal(InputModal):
         self.persons = get_all_persons()
         self.accounts = get_all_accounts_with_balance()
         self.date = date
+        self.shift_pressed = False
 
     # -------------- Helpers ------------- #
 
@@ -127,7 +133,29 @@ class RecordModal(InputModal):
         # set heldValue for the AutoComplete's input
         event.input.heldValue = person.id
 
-    # def on_auto
+    def on_auto_complete_selected(self, event: AutoComplete.Selected) -> None:
+        if "field-label" in event.input.id:
+            template = get_template_by_id(event.input.heldValue)
+            for field in self.form.fields[1:-1]:
+                has_heldValue = field.type in ["autocomplete", "hidden"]
+                fieldWidget = self.query_one(f"#field-{field.key}")
+                if not has_heldValue:
+                    fieldWidget.value = str(getattr(template, field.key))
+                else:
+                    print(fieldWidget, field.key)
+                    fieldWidget.heldValue = getattr(template, field.key)
+                    if "Id" in field.key:
+                        fieldWidget.value = str(
+                            getattr(
+                                getattr(template, field.key.replace("Id", "")), "name"
+                            )
+                        )
+            date = format_date_to_readable(self.form.fields[-1].default_value)
+            type = (
+                "Income" if self.query_one("#field-isIncome").heldValue else "Expense"
+            )
+            account_name = self.query_one("#field-accountId").value
+            super().set_title(f"New {type} on {account_name} for {date}")
 
     # ------------- Callbacks ------------ #
 
@@ -166,12 +194,23 @@ class RecordModal(InputModal):
                 self.splitForm.fields.pop()
             self.splitCount -= 1
 
+    def action_submit_and_template(self) -> None:
+        """Handle shift+enter submission"""
+        self.shift_pressed = True
+        self.action_submit()
+
     def action_submit(self):
         resultRecordForm, errors, isValid = validateForm(self, self.form)
         resultSplitForm, errorsSplit, isValidSplit = validateForm(self, self.splitForm)
         if isValid and isValidSplit:
             resultSplits = self._get_splits_from_result(resultSplitForm)
-            self.dismiss({"record": resultRecordForm, "splits": resultSplits})
+            self.dismiss(
+                {
+                    "record": resultRecordForm,
+                    "splits": resultSplits,
+                    "createTemplate": self.shift_pressed,
+                }
+            )
             return
         self._update_errors({**errors, **errorsSplit})